Rename aggregator subpackages.  Remove date compare logic for kafka messages.  Add create and update methods for entities.
