+++ src/main/java/com/deloitte/elrr/aggregator/consumer/package-info.java
@@ -6,8 +6,6 @@
 
 /**
  * @author mnelakurti
- *
- *This package info file for package com.deloitte.elrr
- *
+ *     <p>This package info file for package com.deloitte.elrr
  */
-package com.deloitte.elrr.elrrconsolidate.dto;
+package com.deloitte.elrr.aggregator.consumer;
