+++ src/main/java/com/deloitte/elrr/aggregator/drools/package-info.java
@@ -1,0 +1,1 @@
+package com.deloitte.elrr.aggregator.drools;
