--- src/main/java/com/deloitte/elrr/drools/package-info.java
@@ -1,1 +1,0 @@
-package com.deloitte.elrr.drools;
